requirements.txt
flask
pandas
requests
python-dotenv
google-genai
google-generativeai
